fix release container build

+++ docker/multi-arch.sh
@@ -5,7 +5,9 @@ set -exo pipefail
 src_dir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 cd "$src_dir"
 
+docker="$(command -v docker 2> /dev/null || echo "podman")"
 GH_REPO="ghcr.io/trow-registry/trow"
+
 
 # Check if cargo-sqlx is installed
 if ! command -v cargo-sqlx >/dev/null 2>&1; then
@@ -35,15 +37,18 @@ else
     echo "→ Development database already exists at $DB_PATH"
 fi
 
-# Use trow-multi builder if it exists, otherwise create it
-if ! docker buildx ls | grep -s trow-multi ;
-then
-    # # Register binfmt handlers
-    docker run --rm --privileged aptman/qus -s -- -p arm aarch64
-    # Create new build instance
-    docker buildx create --name trow-multi
+if [ "$docker" = "docker" ]; then
+    # Use trow-multi builder if it exists, otherwise create it
+    # Not needed for podman
+    if ! docker buildx ls | grep -s trow-multi ;
+    then
+        # # Register binfmt handlers
+        docker run --rm --privileged aptman/qus -s -- -p arm aarch64
+        # Create new build instance
+        docker buildx create --name trow-multi
+    fi
+    docker buildx use trow-multi
 fi
-docker buildx use trow-multi
 
 VERSION=$(sed '/^version = */!d; s///;q' ../Cargo.toml | sed s/\"//g)
 TAG="$VERSION"
@@ -57,7 +62,7 @@ fi
 
 echo $PUSH $GH_IMAGE $GH_REPO
 # Can't load the image in the local registry... See https://github.com/docker/roadmap/issues/371
-docker buildx build \
+$docker buildx build \
   --build-arg VCS_REF="${SOURCE_COMMIT:-$(git rev-parse HEAD)}" \
   --build-arg VCS_BRANCH="${SOURCE_BRANCH:-$(git symbolic-ref --short HEAD)}" \
   --build-arg REPO="$GH_REPO" \
